Rewrote IP fetching.
Needs testing.
